Organize codemirror files
