+++ codemirror/booleanExtension.js
@@ -4,9 +4,9 @@ import {
     WidgetType,
     Decoration,
     ViewPlugin
-} from "./pkg/@codemirror/view.js";
-import { syntaxTree } from "./pkg/@codemirror/language.js";
-import { dispatch } from "./dispatch.js";
+} from "../pkg/@codemirror/view.js";
+import { syntaxTree } from "../pkg/@codemirror/language.js";
+import { dispatch } from "../dispatch.js";
 
 class CheckboxWidget extends WidgetType {
     constructor(checked) {
