+++ codemirror/cm.js
@@ -2,10 +2,10 @@ import {
     EditorState,
     EditorView,
     basicSetup
-} from "./pkg/@codemirror/basic-setup.js";
-import { javascript } from "./pkg/@codemirror/lang-javascript.js";
-import booleanExt from "./cmBooleanExtension.js"
-import spriteExt from "./cmSpriteExtension.js";
+} from "../pkg/@codemirror/basic-setup.js";
+import { javascript } from "../pkg/@codemirror/lang-javascript.js";
+import booleanExt from "./booleanExtension.js"
+import spriteExt from "./spriteExtension.js";
 
 export default function init() {
     return new EditorView({
